+++ src/pages/Application/Application.tsx
@@ -1,44 +1,13 @@
-import type { Application as IApplication } from "@better-giving/registration/approval";
-import { useLoaderData } from "@remix-run/react";
-import type { LoaderFunction, MetaFunction } from "@vercel/remix";
-import { ap, ver } from "api/api";
-import { cognito, redirectToAuth } from "auth";
-import { metas } from "helpers/seo";
+import { useCachedLoaderData } from "api/cache";
 import { CircleAlert } from "lucide-react";
-import type { UserV2 } from "types/auth";
 import Loaded from "./Loaded";
+import type { LoaderData } from "./api";
 
-export interface LoaderData {
-  user: UserV2;
-  application: IApplication;
-}
-
-export const loader: LoaderFunction = async ({ params, request }) => {
-  const { user, headers } = await cognito.retrieve(request);
-  if (!user) return redirectToAuth(request, headers);
-
-  const application = await ap
-    .get<IApplication>(`${ver(1)}/registrations/${params.id}`, {
-      headers: { authorization: user.idToken },
-    })
-    .json();
-
-  return {
-    application,
-    user,
-  } satisfies LoaderData;
-};
-
-export const meta: MetaFunction = ({ data }) => {
-  if (!data) return [];
-  return metas({
-    title: `Application Review - ${(data as LoaderData).application.contact.org_name}`,
-  });
-};
-
+export { meta, loader } from "./api";
+export { clientLoader } from "api/cache";
 export { ErrorBoundary } from "components/error";
 export default function Application() {
-  const { application, user } = useLoaderData() as LoaderData;
+  const { application, user } = useCachedLoaderData() as LoaderData;
 
   if (!user.groups.includes("ap-admin")) {
     return (
@@ -54,7 +23,6 @@ export default function Application() {
       <h1 className="text-center text-3xl col-span-full max-lg:mb-4">
         Applications Review - Details
       </h1>
-
       <Loaded {...application} />
     </div>
   );
